backend/requirements.txt
# --- API Framework ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# --- Settings (đọc .env) ---
pydantic-settings>=2.0.0

# --- Data & Model ---
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.26.0

# --- HTTP ---
requests>=2.31.0
httpx>=0.27.0

# --- Database ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# --- Environment Variables ---
python-dotenv>=1.0.0
